backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
psycopg2-binary==2.9.9
celery==5.3.4
redis==5.0.1
django-cors-headers==4.3.0
dj-database-url==2.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
whitenoise==6.6.0
drf-yasg==1.21.7
twilio==8.9.1
django-celery-beat==2.5.0
django-filter==23.3
Pillow==10.1.0
flower==2.0.1
# Dépendances NLP pour la classification des feedbacks
numpy==1.26.0
scipy==1.11.3
scikit-learn==1.3.2
pandas==2.1.1
nltk==3.8.1
